Added Svg support for SymbolStyle (#322)
* Added Geometry Circle
Added Circle to Skia and WPF renderer, but only Skia works. There are problems with the transformation on WPF, that I couldn't solve.

* Added Svg support for SymbolStyle
Added support for Svg instead of bitmap for SymbolStyle. Is implemented for Skia fully working and drawing in vector style. No conversion to bitmap. For WPF, only the functions are capable of managing this. It's missing a function to convert Svg images in bitmap images in SymbolCache.cs, line 23. Could be made by Skia. If WPF should draw Svg in vector style too, than there must be more coded changed, because SymbolCache.GetOrCreate() has return type BitmapImage.
Added a relative option to Offset. Up to now, Offset was absolut in pixel. With relative option, Offset could be from -0.5 to +0.5 in both directions and is multiplied with the image width/heigth when using. This should work on Skia and WPF. Hope I found all occurences ;)

* Added Svg support for SymbolStyle
Added support for Svg instead of bitmap for SymbolStyle. Is implemented for Skia fully working and drawing in vector style. No conversion to bitmap. For WPF, only the functions are capable of managing this. It's missing a function to convert Svg images in bitmap images in SymbolCache.cs, line 23. Could be made by Skia. If WPF should draw Svg in vector style too, than there must be more coded changed, because SymbolCache.GetOrCreate() has return type BitmapImage.
Added a relative option to Offset. Up to now, Offset was absolut in pixel. With relative option, Offset could be from -0.5 to +0.5 in both directions and is multiplied with the image width/heigth when using. This should work on Skia and WPF. Hope I found all occurences ;)

* Added Svg2Xaml and enabled with this Svg on WPF too

* Added a sample for displaying 200 tigers on the same time

* Added some time stoppers

* Added the tiger as png and a test

* Changed size of tiger.png

* Added test for smaller symbols

* Removed a bug

* Removed parts from Circle PR

* Use bitmapInfo width and height rather than the Svg Canvas properties