Removed ZoomTo method.
Why: The method called withing this method was not
intended for this purpose. It ony worked as expected
with duration set to 0 (and worked differently when
duration was set to > 0) because there was a bug in the
implementation where the 0 duration used the
_animationZoomCenter instead.

The NavigateTo(center, resolution...) should be used instead.