Changing variables for TileLayer from private to protected
The reason is, that with this changes, it is possible to derive from TileLayer and use it not only for raster tiles, but for all other types of tile oriented features. The main things are set via arguments in the constructor and up to now, it isn't possible to replace this constructor, because all variables are private. When making them protected, this would be possible.